Separate videoPane from MonitorScene
